+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/oauth/feign/NaverOAuthFeignClient.java
@@ -1,6 +1,6 @@
-package com.pawith.authapplication.service.command.feign;
+package com.pawith.authapplication.service.oauth.feign;
 
-import com.pawith.authapplication.service.command.feign.response.NaverUserInfo;
+import com.pawith.authapplication.service.oauth.feign.response.NaverUserInfo;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
