+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/oauth/impl/KakaoOAuthHandler.java
@@ -1,10 +1,10 @@
-package com.pawith.authapplication.service.command.handler.impl;
+package com.pawith.authapplication.service.oauth.impl;
 
 import com.pawith.authapplication.dto.OAuthRequest;
 import com.pawith.authapplication.dto.OAuthUserInfo;
-import com.pawith.authapplication.service.command.feign.KakaoOAuthFeignClient;
-import com.pawith.authapplication.service.command.feign.response.KakaoUserInfo;
-import com.pawith.authapplication.service.command.feign.response.TokenInfo;
+import com.pawith.authapplication.service.oauth.feign.KakaoOAuthFeignClient;
+import com.pawith.authapplication.service.oauth.feign.response.KakaoUserInfo;
+import com.pawith.authapplication.service.oauth.feign.response.TokenInfo;
 import com.pawith.authapplication.service.command.handler.AuthHandler;
 import com.pawith.authdomain.exception.AuthError;
 import com.pawith.authdomain.jwt.exception.InvalidTokenException;
